+++ acceptance/tests/login.ts
@@ -1,4 +1,7 @@
-import {Page} from "@playwright/test";
+import {expect, Page} from "@playwright/test";
+import {loginnameScreen} from "./loginname";
+import {passwordScreen} from "./password";
+import {passkeyScreen} from "./passkey";
 
 export async function loginWithPassword(page: Page, username: string, password: string) {
     await page.goto("/loginname");
@@ -8,17 +11,13 @@ export async function loginWithPassword(page: Page, username: string, password: 
     await page.getByTestId("submit-button").click()
 }
 
-export async function loginnameScreen(page: Page, username: string) {
-    await page.getByTestId("username-text-input").pressSequentially(username);
-}
-
-export async function passwordScreen(page: Page, password: string) {
-    await page.getByTestId("password-text-input").pressSequentially(password);
-}
-
 export async function loginWithPasskey(page: Page, username: string) {
     await page.goto("/loginname");
     await loginnameScreen(page, username)
     await page.getByTestId("submit-button").click()
-    await page.getByTestId("submit-button").click()
+    await passkeyScreen(page)
+}
+
+export async function checkLogin(page: Page, fullName: string) {
+    await expect(page.getByRole('heading')).toContainText(fullName);
 }
