+++ acceptance/tests/loginname.ts
@@ -1,0 +1,5 @@
+import {Page} from "@playwright/test";
+
+export async function loginnameScreen(page: Page, username: string) {
+    await page.getByTestId("username-text-input").pressSequentially(username);
+}
