+++ acceptance/tests/password.ts
@@ -10,3 +10,12 @@ async function changePasswordScreen(page: Page, password1: string, password2: st
     await page.getByTestId('password-text-input').pressSequentially(password1);
     await page.getByTestId('password-confirm-text-input').pressSequentially(password2);
 }
+
+export async function passwordScreen(page: Page, password: string) {
+    await page.getByTestId("password-text-input").pressSequentially(password);
+}
+
+export async function registerPasswordScreen(page: Page, password1: string, password2: string) {
+    await page.getByTestId('password-text-input').pressSequentially(password1);
+    await page.getByTestId('password-confirm-text-input').pressSequentially(password2);
+}
