+++ acceptance/tests/register.ts
@@ -1,4 +1,6 @@
 import {Page} from "@playwright/test";
+import {passkeyScreen} from './passkey';
+import {registerPasswordScreen} from './password';
 
 export async function registerWithPassword(page: Page, firstname: string, lastname: string, email: string, password1: string, password2: string) {
     await page.goto('/register');
@@ -10,24 +12,19 @@ export async function registerWithPassword(page: Page, firstname: string, lastna
 
 async function registerUserScreenPassword(page: Page, firstname: string, lastname: string, email: string) {
     await registerUserScreen(page, firstname, lastname, email)
-    await page.getByLabel('Password').click();
-}
-
-async function registerPasswordScreen(page: Page, password1: string, password2: string) {
-    await page.getByTestId('password-text-input').fill(password1);
-    await page.getByTestId('password-confirm-text-input').fill(password2);
+    await page.getByTestId('Password-radio').click();
 }
 
 export async function registerWithPasskey(page: Page, firstname: string, lastname: string, email: string) {
     await page.goto('/register');
     await registerUserScreenPasskey(page, firstname, lastname, email)
     await page.getByTestId('submit-button').click();
-    await page.getByTestId('submit-button').click();
+    await passkeyScreen(page)
 }
 
 async function registerUserScreenPasskey(page: Page, firstname: string, lastname: string, email: string) {
     await registerUserScreen(page, firstname, lastname, email)
-    await page.getByLabel('Passkey').click();
+    await page.getByTestId('Passkeys-radio').click();
 }
 
 async function registerUserScreen(page: Page, firstname: string, lastname: string, email: string) {
