+++ acceptance/tests/username-password-changed.spec.ts
@@ -2,6 +2,7 @@ import {test as base} from "@playwright/test";
 import {PasswordUser} from './user';
 import path from 'path';
 import dotenv from 'dotenv';
+import {checkLogin} from "./login";
 
 // Read from ".env" file.
 dotenv.config({path: path.resolve(__dirname, '.env.local')});
@@ -23,4 +24,5 @@ const test = base.extend<{ user: PasswordUser }>({
 test("username and password changed login", async ({user, page}) => {
     await user.changePassword(page, "ChangedPw1!")
     await user.login(page)
+    await checkLogin(page, user.fullName());
 });
